sfdbs

+++ src/routes/quizzes/deleteQuizById.js
@@ -1,3 +1,4 @@
+import * as admin from 'firebase-admin'
 import { db } from "../../database";
 
 export const deleteQuizByIdRoute = {
@@ -5,9 +6,14 @@ export const deleteQuizByIdRoute = {
     path: '/api/quizzes/{id}',
     handler: async (req, h) => {
         const { id } = req.params;
+        const token = req.headers.authtoken;
+        const user = await admin.auth().verifyIdToken(token);
+        const userId = user.user_id;
+        console.log(id);
+        console.log(userId);
         await db.query(
-            'DELETE FROM quizzes WHERE id=?',
-            [id],
+            'DELETE FROM quizzes WHERE id=? AND user_id=?',
+            [id, userId],
         );
         return { message: 'Success!'}
     }
